RW/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.29.0
flask-jwt-extended==4.5.3
bcrypt==4.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
alembic==1.12.0
jupyterlab==4.0.7
